feat: factor out update_GameScreen() to game_rl.c
chore: handle floor regen by passing double pointers to update_GameScreen()

feat: factor out draw_GameScreen_Texture() to game_rl.c
